Added Checkpoint System
and some fixes

+++ Assets/Scripts/UI/ResultData.cs
@@ -8,12 +8,12 @@ public class ResultData
     
     public static int timeBonusPerSecond = 0;
     public static int maxTime = 0;
-    public static int elapsedTime = 0;
+    public static float elapsedTime = 0;
     public static int score = 0;
     public static int[] rankScoreCutoff;
 
     public static int GetTimeBonusScore()
     {
-        return (maxTime - elapsedTime) * timeBonusPerSecond;
+        return (maxTime - (int)elapsedTime) * timeBonusPerSecond;
     }
 }
